ntos: ob: Implement directory objects + root obj
This commit introduces a root object which exists at path '/'. The root
object is always of a directory type and stores references every other
object and object subdirectory in the namespace. Along with this, this
commit also implements the initialization of the object store as well as
path parsing logic for object lookups.

Signed-off-by: Ian Moffett <ian@osmora.org>