ntos/amd64: Adjust IRQLs for proper serialization
There is a maximum of 15 Interrupt Request Levels, bits 7:4 of the
interrupt vector number is the interrupt priority level (IPL) and bits 3:0 of
the TPR is called the task priority level (TPL). An interrupt with a higher
IPL than the TPL can preempt the current task, however, no events can
preempt a task with a TPL value of 0xF and therefore all interrupts are
deferred. This commit reorders the list of IRQLs to match these
semantics so that there is a clear seperation between lower priority
interrupts and higher priority interrupts as well as giving the guarante
that serializing to HIGH_LEVEL defers all interrupts.

Signed-off-by: Ian Moffett <ian@osmora.org>